Splitted the big TestHelper in three classes.

+++ CK.Env.Sys/FileHelper.cs
@@ -10,58 +10,34 @@ namespace CK.Env
     public static class FileHelper
     {
         /// <summary>
-        /// Helper that deletes a local directory (with linear timed retries).
+        /// Helper that deletes a local directory with retries.
         /// Throws the exception after 4 unsuccessful retries.
-        /// </summary>
-        /// <param name="m">The monitor to use.</param>
-        /// <param name="dirPath">The directory path on the local file system to delete.</param>
-        /// <returns>False if the directory did not exist, true if it has actually been deleted.</returns>
-        public static bool RawDeleteLocalDirectory( IActivityMonitor m, string dirPath )
-        {
-            if( !Directory.Exists( dirPath ) ) return true;
-            m.Info( $"Deleting {dirPath}." );
-            int tryCount = 0;
-            for(; ; )
-            {
-                try
-                {
-                    if( Directory.Exists( dirPath ) ) Directory.Delete( dirPath, true );
-                    return true;
-                }
-                catch( Exception ex )
-                {
-                    m.Warn( $"Error while deleting {dirPath}.", ex );
-                    if( ++tryCount > 4 ) throw;
-                    System.Threading.Thread.Sleep( 100 * tryCount );
-                }
-            }
-        }
-
-        /// <summary>
         /// From https://github.com/libgit2/libgit2sharp/blob/f8e2d42ed9051fa5a5348c1a13d006f0cc069bc7/LibGit2Sharp.Tests/TestHelpers/DirectoryHelper.cs#L40
         /// </summary>
-        /// <param name="directoryPath"></param>
-        public static void DeleteDirectory( IActivityMonitor m, string directoryPath )
+        /// <param name="m">The monitor to use.</param>
+        /// <param name="directoryPath">The directory path on the local file system to delete.</param>
+        /// <returns>True if the Directory was deleted or did not exist, false if it didn't deleted the directory.</returns>
+        public static bool RawDeleteLocalDirectory( IActivityMonitor m, string directoryPath )
         {
             // From http://stackoverflow.com/questions/329355/cannot-delete-directory-with-directory-deletepath-true/329502#329502
 
             if( !Directory.Exists( directoryPath ) )
             {
-                m.Trace( string.Format( "Directory '{0}' is missing and can't be removed.", directoryPath ) );
-                return;
+                m.Trace( $"Directory '{directoryPath}' does not exist." );
+                return true;
             }
             NormalizeAttributes( directoryPath );
-            DeleteDirectory( m, directoryPath, maxAttempts: 5, initialTimeout: 16, timeoutFactor: 2 );
+            return DeleteDirectory( m, directoryPath, maxAttempts: 5, initialTimeout: 16, timeoutFactor: 2 );
         }
         static readonly Type[] _whitelist = { typeof( IOException ), typeof( UnauthorizedAccessException ) };
-        static void DeleteDirectory( IActivityMonitor m, string directoryPath, int maxAttempts, int initialTimeout, int timeoutFactor )
+        static bool DeleteDirectory( IActivityMonitor m, string directoryPath, int maxAttempts, int initialTimeout, int timeoutFactor )
         {
             for( int attempt = 1; attempt <= maxAttempts; attempt++ )
             {
                 try
                 {
-                    Directory.Delete( directoryPath, true );
-                    return;
+                    if( Directory.Exists( directoryPath ) ) Directory.Delete( directoryPath, true );
+                    return true;
                 }
                 catch( Exception ex )
                 {
@@ -85,6 +61,45 @@ namespace CK.Env
                                                   "{0}- Antivirus (exclude the bin folder of LibGit2Sharp.Tests from the paths scanned by your real-time antivirus)" +
                                                   "{0}- TortoiseGit (change the 'Icon Overlays' settings, e.g., adding the bin folder of LibGit2Sharp.Tests to 'Exclude paths' and appending an '*' to exclude all subfolders as well)",
                         Environment.NewLine, Path.GetFullPath( directoryPath ), maxAttempts, caughtExceptionType, ex.Message ) );
+                }
+            }
+            return false;
+        }
+
+        public static void DirectoryCopy( string sourceDirName, string destDirName, bool copySubDirs )
+        {
+            // Get the subdirectories for the specified directory.
+            DirectoryInfo dir = new DirectoryInfo( sourceDirName );
+
+            if( !dir.Exists )
+            {
+                throw new DirectoryNotFoundException(
+                    "Source directory does not exist or could not be found: "
+                    + sourceDirName );
+            }
+
+            DirectoryInfo[] dirs = dir.GetDirectories();
+            // If the destination directory doesn't exist, create it.
+            if( !Directory.Exists( destDirName ) )
+            {
+                Directory.CreateDirectory( destDirName );
+            }
+
+            // Get the files in the directory and copy them to the new location.
+            FileInfo[] files = dir.GetFiles();
+            foreach( FileInfo file in files )
+            {
+                string temppath = Path.Combine( destDirName, file.Name );
+                file.CopyTo( temppath, false );
+            }
+
+            // If copying subdirectories, copy them and their contents to new location.
+            if( copySubDirs )
+            {
+                foreach( DirectoryInfo subdir in dirs )
+                {
+                    string temppath = Path.Combine( destDirName, subdir.Name );
+                    DirectoryCopy( subdir.FullName, temppath, copySubDirs );
                 }
             }
         }
